Extract limit orders widget from page layout
